// apps/web/src/characterBuilder/generated/backgroundsLookup.ts
export const backgroundsLookup: BackgroundsLookup = {
  "byId": {
    "acolyte": {
      "id": "acolyte",
      "slug": "acolyte",
      "name": "Acolyte",
      "aliases": [
        "Acolyte"
      ],
      "categories": [
        "Common Backgrounds"
      ],
      "summary": "You have spent your life in the service of a temple to a specific god or pantheon of gods.",
      "grants": {
        "skills": [
          "Insight",
          "Religion"
        ],
        "tools": [],
        "languages": [],
        "skillChoices": null,
        "toolChoices": null,
        "languageChoices": {
          "choose": 2,
          "from": [
            "Common",
            "Dwarvish",
            "Elvish",
            "Giant",
            "Gnomish",
            "Goblin",
            "Halfling",
            "Orc",
            "Abyssal",
            "Celestial",
            "Draconic",
            "Deep Speech",
            "Infernal",
            "Primordial",
            "Sylvan",
            "Undercommon"
          ]
        }
      },
      "featureName": "Shelter of the Faithful",
      "equipmentSummary": [
        "holy symbol (a gift to you when you entered the priesthood)",
        "5x sticks of incense",
        "vestments",
        "set of common clothes",
        "pouch containing 15gp"
      ],
      "detailUrl": "/rules/backgrounds/acolyte.json"
    },
// ...
  },
// ...
};